refactor: extract Broadcastable, Notifiable, Approvable concerns from Comment model
- Comment model reduced from 345 to 92 lines (-253 lines)
- Comment::Broadcastable: Turbo broadcast + badge logic (incl. class methods)
- Comment::Notifiable: inbox notifications, mentions, approver alerts
- Comment::Approvable: approval_status, can_be_approved_by?
- All 8 existing model tests + 13 Creative model tests pass
